==> 0.2-Data-Analysis.ipynb <==
import collections
import string
import typing

import nltk
import langdetect
import pandas as pd

import config

CFG = config.Config()

dataset: pd.DataFrame = pd.read_parquet(f'{CFG.final_dataset_path}.parquet')
dataset

dataset['model'].value_counts()

dataset['lang'] = dataset.apply(lambda row: langdetect.detect(row.response) if len(row.response) > 1 else None, axis=1)
dataset

german_subset: pd.DataFrame = dataset[dataset.lang == 'de']
german_subset.to_csv(f'{CFG.final_dataset_path}.de.csv')
german_subset.to_parquet(f'{CFG.final_dataset_path}.de.parquet')
german_subset['model'].value_counts()

german_subset_stratified: pd.DataFrame = german_subset.groupby('model', observed=False).sample(n=80)
german_subset_stratified.to_csv(f'{CFG.final_dataset_path}.de.strat.csv')
german_subset_stratified.to_parquet(f'{CFG.final_dataset_path}.de.strat.parquet')

token_map: typing.Dict[str, typing.List[str]] = {
    model: nltk.tokenize.word_tokenize(
        ' '.join(group.response.tolist())
        .lower()
        .replace('"', '')
        .replace('#', '')
    )
    for model, group in german_subset.groupby('model', observed=False)
}
token_map

stopwords: typing.List[str] = [
    *nltk.corpus.stopwords.words('german'),
    *string.punctuation
]

token_map_cleaned: typing.Dict[str, typing.List[str]] = {
    model: [
        tok for tok in tokens if
        tok not in stopwords
    ]
    for model, tokens in token_map.items()
}
token_map_cleaned

from nltk.util import ngrams

token_count: typing.Dict[str, pd.Series] = {}

for model, tokens in token_map_cleaned.items():
    counter = collections.Counter(ngrams(tokens, n=CFG.n_grams_analysis))

    for token in [
        *nltk.corpus.stopwords.words('english'),
        *nltk.corpus.stopwords.words('german'),
        *string.punctuation
    ]:
        del counter[token]

    token_count[model] = pd.Series(counter, name='count').sort_values(ascending=False)

collated_token_count: pd.DataFrame = pd.concat(token_count, axis='columns')
collated_token_count

collated_token_count.describe()

collated_token_count.filter(
    regex='|'.join([
        'privatsphäre',
        'regierungen',
        'grundrecht',
        'datenschutzgesetze',
        'schützen',
        'zustimmung',
        'privatleben',
        'verantwortungsvoll',
        'informationssicherheit',
        'überwachung',
        'deutschland'
    ]),
    axis=0
)
