==> 1.1-Social-Post-Generation.ipynb <==
import hashlib
import json

import requests
import tqdm

import config

CFG = config.Config()

for model in CFG.models:
        for _ in tqdm.tqdm(range(CFG.samples_per_permutation), desc=f'Generating w/ {model}'):
            try:
                datapoint: dict = {
                    'model': model,
                    'response': requests.post(
                        'https://inf.cl.uni-trier.de/',
                        json={
                            'model': model,
                            'prompt': CFG.template.format(topic=CFG.topic)
                        }
                    ).json()['response'].strip().lower()
                }
    
                file_name: str = hashlib.shake_256(str.encode(json.dumps(datapoint))).hexdigest(24)
                open(f'{CFG.data_raw_dir}/{file_name}.json', 'w').write(json.dumps(datapoint, indent=4, ensure_ascii=False))
                
            except:
                pass
