==> 1.3-Human-Tweet-Collecting.ipynb <==
import itertools

import pandas as pd
import tweety as twy

import config

CFG = config.Config()

LOGIN_USER: str = "***********"
LOGIN_PWD: str = "***********"

PAGES: int = 5
KEYWORDS: list = ['datenschutz', 'datensicherheit']

app = twy.Twitter("session")
app.sign_in(LOGIN_USER, LOGIN_PWD)

posts_sample: pd.DataFrame = pd.concat([
    pd.json_normalize([
        tweet for tweet in app.search(keyword, pages=page)
    ])
    for page, keyword in itertools.product(*[range(1, PAGES + 1), KEYWORDS])
])
posts_sample.to_json(f'{CFG.data_dir}/raw.human.json', orient='records', indent=4)

posts_sample
