==> 2.3-Human-Tweet-Processing.ipynb <==
import pandas as pd

import config

CFG = config.Config()

raw: pd.DataFrame = pd.read_json(f'{CFG.data_dir}/raw.human.json')
raw

processed: pd.DataFrame = (
    raw
    .set_index('id', drop=True)
    [['text']]
    .assign(model='human')
    .assign(lang=CFG.lang)
    .rename(columns={'text': 'response'})
)
processed

processed.to_csv(f'{CFG.data_dir}/dataset.human.{CFG.lang}.csv')
processed.to_parquet(f'{CFG.data_dir}/dataset.human.{CFG.lang}.parquet')

dataset_prod = (
    pd.concat(
        [
            (synthetic := pd.read_parquet(f'{CFG.data_dir}/{CFG.dataset_name}.{CFG.lang}.strat.parquet')),
            processed.sample(synthetic.value_counts('model').min())
        ]
    )
    .reset_index(drop=True)
)
dataset_prod.to_csv(f'{CFG.data_dir}/dataset.prod.{CFG.lang}.csv')
dataset_prod.to_parquet(f'{CFG.data_dir}/dataset.prod.{CFG.lang}.parquet')
