==> 3.1-Argument-Mining.ipynb <==
import pandas as pd
import requests

import config

CFG = config.Config()

PROMPT: str = \
"""
You are an argument miner and your purpose is retrieving arguments and topics from social media post and present them structured as a response.
"""

data: pd.DataFrame = pd.read_parquet(f'{CFG.data_dir}/{CFG.dataset_name}.prod.{CFG.lang}.parquet')
data

for name, group in data.groupby('model'):
    if name == 'mixtral':
        (
            open(f'{CFG.data_dir}/argument_evaluation.{name}.{CFG.lang}.txt', 'w')
            .write(
                requests.post(
                'https://inf.cl.uni-trier.de/',
                json={
                    'model': 'llama3:70b-instruct-q6_K',
                    'system': PROMPT,
                    'prompt': f'Posts:\n{group["response"].tolist()}\n\nArguments:\n'
                }
            ).json()['response']
            )
        )
