==> 3.2-Moral-Foundation-Evaluation.ipynb <==
import typing
import json

import pandas as pd

import config

CFG = config.Config()

dataset: pd.DataFrame = (
    pd.read_parquet(f'{CFG.data_dir}/questionary.responses.parquet')
    .astype(pd.Int64Dtype())
)
dataset

mapping = json.loads(open('questionary_mapping.json').read())
mapping

evaluation = pd.concat([(
    pd.concat(
        [dataset.loc[tuple(quest)] for quest in questions],
        axis=1)
    .mean(axis=1)
    .unstack()
    .rename(index={'response': label})
)
    for label, questions in mapping.items()]).T

evaluation.to_parquet(f'{CFG.data_dir}/questionary.evaluation.parquet')
evaluation.to_latex(f'{CFG.report_dir}/questionary.evaluation.tex', float_format=lambda f: f'{f:2.3f}')
evaluation

human_results = {
    "anon": {
        "liberal": {"authority": 2.2, "purity": 2.2, "ingroup":  2.5, "harm": 4.25, "fairness": 4.35},
        "moderate": {"authority": 2.8, "purity": 3.0, "ingroup":  3.0, "harm": 4.0, "fairness": 3.95},
        "conservative": {"authority": 3.1, "purity": 3.4 , "ingroup":  3.12, "harm": 3.5, "fairness": 3.65},
    },
    "us": {
        "liberal": {"authority": 1.9, "purity": 1.2, "ingroup":  2.05, "harm": 3.6, "fairness": 3.75},
        "moderate": {"authority": 2.65, "purity": 2.2, "ingroup":  2.65, "harm": 3.4, "fairness": 3.45},
        "conservative": {"authority": 3.5, "purity": 3.1 , "ingroup":  3.4, "harm": 3.0, "fairness": 3.1},
    },
    "korean": {
        "liberal": {"authority": 2.25, "purity": 2.4, "ingroup":  2.3, "harm": 3.35, "fairness": 3.5},
        "moderate": {"authority": 2.55, "purity": 2.65, "ingroup":  2.65, "harm": 3.25, "fairness": 3.35},
        "conservative": {"authority": 2.9, "purity": 2.75 , "ingroup":  2.8, "harm": 3.35, "fairness": 3.1},
    }
}

human_cross_evaluation: typing.List[typing.Dict] = []


for group_label, group in human_results.items():
    for human_label, human in group.items():
        for model_label, model in evaluation.iterrows():
            if None in model.to_dict().values():
                human_cross_evaluation.append({
                    "human_label": f'{group_label}_{human_label}', 
                    "model_label": model_label, 
                    "value": None
                })
            else:
                human_cross_evaluation.append({
                    "human_label": f'{group_label}_{human_label}', 
                    "model_label": model_label, 
                    "value": sum(
                        [
                            abs(value - model.to_dict()[keys]) 
                            for keys, value in human.items()
                        ]
                    ) / len(model)
                })

human_eval_df = pd.DataFrame(human_cross_evaluation).pivot(index='model_label', columns='human_label', values='value')
human_eval_df.to_parquet(f'{CFG.data_dir}/questionary.human.cross-evaluation.parquet')
human_eval_df.to_latex(f'{CFG.report_dir}/questionary.human.cross-evaluation.tex', float_format=lambda f: f'{f:2.3f}')
